refactor: associate module
