+++ src/infra/http/nest/modules/associate/associate.module.ts
@@ -1,7 +1,13 @@
 import { Module } from '@nestjs/common';
 
 import { PrismaService } from '@/infra/repositories/prisma/prisma.service';
-import { AssociateController } from '@/infra/http/nest/modules/associate/associate.controller';
+import {
+  CreateAssociateController,
+  DeleteAssociateController,
+  FindAllAssociateController,
+  FindByIdAssociateController,
+  UpdateAssociateController,
+} from '@/infra/http/nest/modules/associate';
 import { CreateAssociateService } from '@/domains/associate/services/create-associate.service';
 import { UpdateAssociateService } from '@/domains/associate/services/update-associate.service';
 import { ListAssociateService } from '@/domains/associate/services/list-associate.service';
@@ -10,7 +16,13 @@ import { DeleteAssociateByIdService } from '@/domains/associate/services/delete-
 import { AssociateRepositoryPrismaAdapter } from '@/infra/repositories/prisma/domains/associate';
 
 @Module({
-  controllers: [AssociateController],
+  controllers: [
+    CreateAssociateController,
+    DeleteAssociateController,
+    FindAllAssociateController,
+    FindByIdAssociateController,
+    UpdateAssociateController,
+  ],
   providers: [
     {
       provide: CreateAssociateService,
