+++ src/infra/http/nest/modules/associate/delete.controller.ts
@@ -1,0 +1,26 @@
+import { z } from 'zod';
+import { ApiTags } from '@nestjs/swagger';
+import { Controller, Param, Delete, UsePipes, HttpCode } from '@nestjs/common';
+
+import { ZodValidationPipe } from '@/infra/http/nest/@config/pipes/zod-validation-pipe';
+
+import { findAssociateByIdSchema } from '@/domains/associate/dto';
+import { DeleteAssociateByIdService } from '@/domains/associate/services/delete-associate-by-id.service';
+
+const associateByIdValidate = new ZodValidationPipe(findAssociateByIdSchema);
+
+type AssociateById = z.infer<typeof findAssociateByIdSchema>;
+@ApiTags('Associates')
+@Controller('associate')
+export class DeleteAssociateController {
+  constructor(
+    private readonly deleteAssociateByIdService: DeleteAssociateByIdService,
+  ) {}
+
+  @Delete(':id')
+  @UsePipes(associateByIdValidate)
+  @HttpCode(204)
+  remove(@Param(associateByIdValidate) associateById: AssociateById) {
+    return this.deleteAssociateByIdService.execute(associateById.id);
+  }
+}
