+++ src/infra/http/nest/modules/associate/index.ts
@@ -1,0 +1,5 @@
+export * from './create.controller';
+export * from './delete.controller';
+export * from './find-all.controller';
+export * from './find-by-id.controller';
+export * from './update.controller';
